requirements.txt
aiogram==3.4.1
fastapi==0.115.0
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiosqlite==0.19.0
asyncpg==0.29.0
python-multipart==0.0.22
aiofiles==23.2.1
Pillow==10.2.0
